+++ R/crossref.R
@@ -183,7 +183,11 @@ ref_info <- function(paper) {
 #' @return doi
 #' @export
 #' @examples
-#' ref <- "Lakens, D., Mesquida, C., Rasti, S., & Ditroilo, M. (2024). The benefits of preregistration and Registered Reports. Evidence-Based Toxicology, 2(1)."
+#' ref <- paste(
+#'   "Lakens, D., Mesquida, C., Rasti, S., & Ditroilo, M. (2024).",
+#'   "The benefits of preregistration and Registered Reports.",
+#'   "Evidence-Based Toxicology, 2(1)."
+#' )
 #' \donttest{
 #'   doi <- get_doi(ref)
 #' }