+++ tests/testthat/test-02-paper.R
@@ -18,7 +18,7 @@ test_that("paper", {
 
   xml <- demoxml()
   p <- paper(xml)
-  expect_equal(nrow(p$full_text), 27)
+  expect_equal(nrow(p$full_text), 28)
   expect_equal(nrow(p$bib), 4)
   expect_equal(nrow(p$xrefs), 2)
 })